Add delete component
