+++ src/components/navbar/navbar.css
@@ -1,0 +1,36 @@
+.nav-container {
+  display: flex;
+  flex-direction: column;
+  width: 20%;
+  height: 100vh;
+  padding: 0 0 0 12px;
+  border-right: 1px solid var(--text-color-secondary);
+}
+
+.profile {
+  margin-bottom: 120px;
+}
+
+.nav-links {
+  position: relative;
+  display: flex;
+  margin-left: 12px;
+  flex-direction: column;
+}
+
+.nav-links > a {
+  text-decoration: none;
+  color: #000;
+  padding: 12px 0;
+  text-transform: uppercase;
+  font-weight: 700;
+}
+
+.nav-links a.active {
+  background-color: var(--bg-button-color) !important;
+  color: #fff !important;
+}
+
+.logout {
+  margin-top: auto;
+}
